requirements.txt
langchain-community
langchain-ollama
langchain
langsmith
chromadb
pypdf
tqdm
python-dotenv